+++ javascript/UnixLepik.js
@@ -299,6 +299,10 @@ class UnixLepik extends LepikEvents_1.LepikEvents {
         const command = `xdotool windowmove ${windowId} ${x} ${y}`;
         __classPrivateFieldGet(this, _UnixLepik_instances, "m", _UnixLepik_executeShellCommand).call(this, command);
     }
+    focusNextWindow() {
+        const command = `xdotool key --clearmodifiers Alt+Tab`;
+        __classPrivateFieldGet(this, _UnixLepik_instances, "m", _UnixLepik_executeShellCommand).call(this, command);
+    }
     /**
      * Delays the execution for the specified number of milliseconds.
      * @param {number} ms - The number of milliseconds to delay.